Adjust libaugeas.so.0 path hack for snapcraft
When running inside snapcraft-the-snap, SNAP is set at snap build time,
causing this code to run and attempt to find libaugeas.so.0 inside
snapcraft's own snap. This is incorrect; at build time, we can expect
libaugeas.so.0 to exist on the build system.

SNAP_NAME should always be set to 'certbot' at runtime, so we can test
for this to conditionally enable the hack only when required.

This fixes the build when snapcraft is run from inside a snap.